refactor/logger test
