+++ .eslintignore
@@ -4,3 +4,4 @@ test/manager/**/_fixtures
 test/platform/**/_fixtures
 test/config/**/_fixtures
 test/workers/**/_fixtures
+test/logger/**/_fixtures
