+++ .prettierignore
@@ -15,4 +15,5 @@ test/manager/**/_fixtures
 test/platform/**/_fixtures
 test/config/**/_fixtures
 test/workers/**/_fixtures
+test/logger/**/_fixtures
 bin/yarn*
