+++ test/globals.js
@@ -4,7 +4,7 @@ jest.mock('gl-got');
 const cache = require('../lib/workers/global/cache');
 
 global.platform = jest.genMockFromModule('../lib/platform/github');
-global.logger = require('./_fixtures/logger');
+global.logger = require('./logger/_fixtures');
 
 global.renovateUsername = 'renovate-testing';
 global.repoCache = {};
